refactor: 모달 닫기 기능 커스텀 훅으로 분리
