+++ components/src/shared/components/modal/Modal.tsx
@@ -1,6 +1,7 @@
 import { createPortal } from 'react-dom';
 import { ModalContainer, ModalLayout } from './Modal.css';
-import { PropsWithChildren, useEffect, useId } from 'react';
+import { PropsWithChildren, useId } from 'react';
+import useEscapeModal from '../../../hooks/useEscapeModal';
 
 interface ModalProps extends PropsWithChildren {
   closeModal: () => void;
@@ -10,27 +11,16 @@ function Modal(props: ModalProps) {
   const { children, closeModal, ...rest } = props;
   const id = useId();
 
-  function handleClickOverlay(e: React.MouseEvent<HTMLDivElement, MouseEvent>) {
-    if (e.target instanceof HTMLElement && e.target.id === id) {
-      closeModal();
-    }
-  }
-
-  useEffect(() => {
-    function handleKeyDown(e: KeyboardEvent) {
-      if (e.key === 'Escape') {
-        closeModal();
-      }
-    }
-
-    window.addEventListener('keydown', handleKeyDown);
-    return () => window.removeEventListener('keydown', handleKeyDown);
-  }, [closeModal]);
+  const { handleClickOverlay } = useEscapeModal(closeModal);
 
   return (
     <div>
       {createPortal(
-        <div css={ModalLayout} id={id} onClick={handleClickOverlay}>
+        <div
+          css={ModalLayout}
+          id={id}
+          onClick={(e) => handleClickOverlay(e, id)}
+        >
           <div css={ModalContainer} {...rest}>
             {children}
           </div>
